AXON-585: file structure refactoring
