+++ e2e/helpers/auth/index.ts
@@ -1,0 +1,6 @@
+export {
+    authenticateWithJira,
+    authenticateWithBitbucketDC,
+    authenticateWithBitbucketCloud,
+    openAtlassianSettings,
+} from './authHelpers';
