+++ e2e/helpers/frame/index.ts
@@ -1,0 +1,1 @@
+export { getIssueFrame } from './frameHelpers';
