+++ e2e/helpers/types.ts
@@ -1,0 +1,6 @@
+export interface JiraIssue {
+    fields: Record<string, any>;
+    renderedFields: Record<string, any>;
+}
+
+export type FieldUpdater = (issue: JiraIssue, value: string | object) => JiraIssue;
