+++ e2e/helpers/wiremock/index.ts
@@ -1,0 +1,1 @@
+export { cleanupWireMockMapping, setupWireMockMapping } from './mappings';
